changement en uint32_t

+++ MAX30102.c
@@ -87,14 +87,14 @@ void readMultipleBytes(uint8_t baseAddress, uint8_t *buffer, uint8_t length)
         I2C_MAX_MasterSendStop(I2C_TIMEOUT); 
 }
 
-void readFIFO(float32_t *red_LED, float32_t *ir_LED, uint16_t compteur){
+void readFIFO(uint32_t *red_LED, uint32_t *ir_LED, uint16_t compteur){
     //ou remplacer par tableau
     uint8_t i2c_data [6];
         
         readMultipleBytes(REG_FIFO_DATA,i2c_data,6);
-        red_LED[compteur]=(float32_t)((0b00000011&i2c_data[0])<<16)+(i2c_data[1]<<8)+(i2c_data[2]);
-        ir_LED[compteur]=(float32_t) ((i2c_data[3]<<16)+(i2c_data[4]<<8)+(i2c_data[5]));
-        
+        red_LED[compteur]=((0b00000011&i2c_data[0])<<16)+(i2c_data[1]<<8)+(i2c_data[2]);
+        ir_LED[compteur]=((i2c_data[3]<<16)+(i2c_data[4]<<8)+(i2c_data[5]));
+        //changement en uint32_t et enlever le cast en (float)
     
     
 }
